moved Endpoint from canmatrix to own package. Class has no tests.
